revm_context::context

Struct Context

pub struct Context<BLOCK = BlockEnv, TX = TxEnv, CFG = CfgEnv, DB: Database = EmptyDB, JOURNAL: JournalTr<Database = DB> = Journal<DB>, CHAIN = (), LOCAL: LocalContextTr = LocalContext> {
    pub block: BLOCK,
    pub tx: TX,
    pub cfg: CFG,
    pub journaled_state: JOURNAL,
    pub chain: CHAIN,
    pub local: LOCAL,
    pub error: Result<(), ContextError<DB::Error>>,
}

EVM context contains data that EVM needs for execution.

Fields
block: BLOCK

Block information.

tx: TX

Transaction information.

cfg: CFG

Configurations.

journaled_state: JOURNAL

EVM State with journaling support and database.

chain: CHAIN

Inner context.

local: LOCAL

Local context that is filled by execution.

error: Result<(), ContextError<DB::Error>>

Error that happened during execution.

Implementations

impl<BLOCK: Block + Default, TX: Transaction + Default, DB: Database, JOURNAL: JournalTr<Database = DB>, CHAIN: Default> Context<BLOCK, TX, CfgEnv, DB, JOURNAL, CHAIN>

pub fn new(db: DB, spec: SpecId) -> Self

Creates a new context with a new database type.

This will create a new Journal object.

impl<BLOCK, TX, CFG, DB, JOURNAL, CHAIN> Context<BLOCK, TX, CFG, DB, JOURNAL, CHAIN>
where BLOCK: Block, TX: Transaction, CFG: Cfg, DB: Database, JOURNAL: JournalTr<Database = DB>,

pub fn with_new_journal<OJOURNAL: JournalTr<Database = DB>>( self, journal: OJOURNAL, ) -> Context<BLOCK, TX, CFG, DB, OJOURNAL, CHAIN>

Creates a new context with a new journal type. New journal needs to have the same database type.

pub fn with_db<ODB: Database>( self, db: ODB, ) -> Context<BLOCK, TX, CFG, ODB, Journal<ODB>, CHAIN>

Creates a new context with a new database type.

This will create a new Journal object.

pub fn with_ref_db<ODB: DatabaseRef>( self, db: ODB, ) -> Context<BLOCK, TX, CFG, WrapDatabaseRef<ODB>, Journal<WrapDatabaseRef<ODB>>, CHAIN>

Creates a new context with a new DatabaseRef type.

pub fn with_block<OB: Block>( self, block: OB, ) -> Context<OB, TX, CFG, DB, JOURNAL, CHAIN>

Creates a new context with a new block type.

pub fn with_tx<OTX: Transaction>( self, tx: OTX, ) -> Context<BLOCK, OTX, CFG, DB, JOURNAL, CHAIN>

Creates a new context with a new transaction type.

pub fn with_chain<OC>( self, chain: OC, ) -> Context<BLOCK, TX, CFG, DB, JOURNAL, OC>

Creates a new context with a new chain type.

pub fn with_cfg<OCFG: Cfg>( self, cfg: OCFG, ) -> Context<BLOCK, TX, OCFG, DB, JOURNAL, CHAIN>

Creates a new context with a new chain type.

pub fn modify_cfg_chained<F>(self, f: F) -> Self
where F: FnOnce(&mut CFG),

Modifies the context configuration.

pub fn modify_block_chained<F>(self, f: F) -> Self
where F: FnOnce(&mut BLOCK),

Modifies the context block.

pub fn modify_tx_chained<F>(self, f: F) -> Self
where F: FnOnce(&mut TX),

Modifies the context transaction.

pub fn modify_chain_chained<F>(self, f: F) -> Self
where F: FnOnce(&mut CHAIN),

Modifies the context chain.

pub fn modify_db_chained<F>(self, f: F) -> Self
where F: FnOnce(&mut DB),

Modifies the context database.

pub fn modify_journal_chained<F>(self, f: F) -> Self
where F: FnOnce(&mut JOURNAL),

Modifies the context journal.

pub fn modify_block<F>(&mut self, f: F)
where F: FnOnce(&mut BLOCK),

Modifies the context block.

pub fn modify_tx<F>(&mut self, f: F)
where F: FnOnce(&mut TX),

Modifies the context transaction.

pub fn modify_cfg<F>(&mut self, f: F)
where F: FnOnce(&mut CFG),

Modifies the context configuration.

pub fn modify_chain<F>(&mut self, f: F)
where F: FnOnce(&mut CHAIN),

Modifies the context chain.

pub fn modify_db<F>(&mut self, f: F)
where F: FnOnce(&mut DB),

Modifies the context database.

pub fn modify_journal<F>(&mut self, f: F)
where F: FnOnce(&mut JOURNAL),

Modifies the context journal.