revm_context::journal::inner

Struct JournalInner

pub struct JournalInner<ENTRY> {
    pub state: EvmState,
    pub transient_storage: TransientStorage,
    pub logs: Vec<Log>,
    pub depth: usize,
    pub journal: Vec<ENTRY>,
    pub spec: SpecId,
    pub warm_preloaded_addresses: HashSet<Address>,
    pub precompiles: HashSet<Address>,
}

Inner journal state that contains journal and state changes.

Spec Id is a essential information for the Journal.

Fields
state: EvmState

The current state

transient_storage: TransientStorage

Transient storage that is discarded after every transaction.

See EIP-1153.

logs: Vec<Log>

Emitted logs

depth: usize

The current call stack depth

journal: Vec<ENTRY>

The journal of state changes, one for each call

spec: SpecId

The spec ID for the EVM. Spec is required for some journal entries and needs to be set for JournalInner to be functional.

If spec is set it assumed that precompile addresses are set as well for this particular spec.

This spec is used for two things:

EIP-161: Prior to this EIP, Ethereum had separate definitions for empty and non-existing accounts.

EIP-6780: SELFDESTRUCT only in same transaction

warm_preloaded_addresses: HashSet<Address>

Warm loaded addresses are used to check if loaded address should be considered cold or warm loaded when the account is first accessed.

Note that this not include newly loaded accounts, account and storage is considered warm if it is found in the State.

precompiles: HashSet<Address>

Precompile addresses

Implementations

impl<ENTRY: JournalEntryTr> JournalInner<ENTRY>

pub fn new() -> JournalInner<ENTRY>

Creates new JournalInner.

warm_preloaded_addresses is used to determine if address is considered warm loaded. In ordinary case this is precompile or beneficiary.

pub fn clear_and_take_output(&mut self) -> JournalOutput

Take the JournalOutput and clears the journal by resetting it to initial state.

Note: Precompile addresses and spec are preserved and initial state of warm_preloaded_addresses will contain precompiles addresses. Precompile addresses

pub fn state(&mut self) -> &mut EvmState

Return reference to state.

pub fn set_spec_id(&mut self, spec: SpecId)

Sets SpecId.

pub fn touch(&mut self, address: Address)

Mark account as touched as only touched accounts will be added to state. This is especially important for state clear where touched empty accounts needs to be removed from state.

pub fn account(&self, address: Address) -> &Account

Returns the loaded Account for the given address.

This assumes that the account has already been loaded.

Panics

Panics if the account has not been loaded and is missing from the state set.

pub fn set_code_with_hash( &mut self, address: Address, code: Bytecode, hash: B256, )

Set code and its hash to the account.

Note: Assume account is warm and that hash is calculated from code.

pub fn set_code(&mut self, address: Address, code: Bytecode)

Use it only if you know that acc is warm.

Assume account is warm.

In case of EIP-7702 code with zero address, the bytecode will be erased.

pub fn inc_nonce(&mut self, address: Address) -> Option<u64>

Increments the nonce of the account.

Returns

Returns the new nonce if it did not overflow, otherwise returns None.

pub fn transfer<DB: Database>( &mut self, db: &mut DB, from: Address, to: Address, balance: U256, ) -> Result<Option<TransferError>, DB::Error>

Transfers balance from two accounts. Returns error if sender balance is not enough.

pub fn create_account_checkpoint( &mut self, caller: Address, target_address: Address, balance: U256, spec_id: SpecId, ) -> Result<JournalCheckpoint, TransferError>

Creates account or returns false if collision is detected.

There are few steps done:

Make created account warm loaded (AccessList) and this should be done before subroutine checkpoint is created.

Check if there is collision of newly created account with existing one.

Mark created account as created.

Add fund to created account

Increment nonce of created account if SpuriousDragon is active

Decrease balance of caller account.

Panics

Panics if the caller is not loaded inside the EVM state. This should have been done inside create_inner.

pub fn checkpoint(&mut self) -> JournalCheckpoint

Makes a checkpoint that in case of Revert can bring back state to this point.

pub fn checkpoint_commit(&mut self)

Commits the checkpoint.

pub fn checkpoint_revert(&mut self, checkpoint: JournalCheckpoint)

Reverts all changes to state until given checkpoint.

pub fn selfdestruct<DB: Database>( &mut self, db: &mut DB, address: Address, target: Address, ) -> Result<StateLoad<SelfDestructResult>, DB::Error>

Performs selfdestruct action. Transfers balance from address to target. Check if target exist/is_cold

Note: Balance will be lost if address and target are the same BUT when current spec enables Cancun, this happens only when the account associated to address is created in the same tx

References:

https://github.com/ethereum/go-ethereum/blob/141cd425310b503c5678e674a8c3872cf46b7086/core/vm/instructions.go#L832-L833

https://github.com/ethereum/go-ethereum/blob/141cd425310b503c5678e674a8c3872cf46b7086/core/state/statedb.go#L449

https://eips.ethereum.org/EIPS/eip-6780

pub fn load_account<DB: Database>( &mut self, db: &mut DB, address: Address, ) -> Result<StateLoad<&mut Account>, DB::Error>

Loads account into memory. return if it is cold or warm accessed

pub fn initial_account_load<DB: Database>( &mut self, db: &mut DB, address: Address, storage_keys: impl IntoIterator<Item = StorageKey>, ) -> Result<StateLoad<&mut Account>, DB::Error>

๐Deprecated: Use load_account_optional_with_storage instead

Initial load of account. This load will not be tracked inside journal

pub fn load_account_delegated<DB: Database>( &mut self, db: &mut DB, address: Address, ) -> Result<StateLoad<AccountLoad>, DB::Error>

Loads account into memory. If account is EIP-7702 type it will additionally load delegated account.

It will mark both this and delegated account as warm loaded.

Returns information about the account (If it is empty or cold loaded) and if present the information about the delegated account (If it is cold loaded).

pub fn load_code<DB: Database>( &mut self, db: &mut DB, address: Address, ) -> Result<StateLoad<&mut Account>, DB::Error>

Loads account and its code. If account is already loaded it will load its code.

It will mark account as warm loaded. If not existing Database will be queried for data.

In case of EIP-7702 delegated account will not be loaded, Self::load_account_delegated should be used instead.

pub fn load_account_optional<DB: Database>( &mut self, db: &mut DB, address: Address, load_code: bool, ) -> Result<StateLoad<&mut Account>, DB::Error>

Loads account. If account is already loaded it will be marked as warm.

pub fn load_account_optional_with_storage<DB: Database>( &mut self, db: &mut DB, address: Address, load_code: bool, storage_keys: impl IntoIterator<Item = StorageKey>, ) -> Result<StateLoad<&mut Account>, DB::Error>

Loads account. If account is already loaded it will be marked as warm.

pub fn sload<DB: Database>( &mut self, db: &mut DB, address: Address, key: StorageKey, ) -> Result<StateLoad<StorageValue>, DB::Error>

Loads storage slot.

Panics

Panics if the account is not present in the state.

pub fn sstore<DB: Database>( &mut self, db: &mut DB, address: Address, key: StorageKey, new: StorageValue, ) -> Result<StateLoad<SStoreResult>, DB::Error>

Stores storage slot.

And returns (original,present,new) slot value.

Note: Account should already be present in our state.

pub fn tload(&mut self, address: Address, key: StorageKey) -> StorageValue

Read transient storage tied to the account.

EIP-1153: Transient storage opcodes

pub fn tstore(&mut self, address: Address, key: StorageKey, new: StorageValue)

Store transient storage tied to the account.

If values is different add entry to the journal so that old state can be reverted if that action is needed.

EIP-1153: Transient storage opcodes

pub fn log(&mut self, log: Log)

Pushes log into subroutine.