revm_context::result

Enum InvalidTransaction

pub enum InvalidTransaction {

    PriorityFeeGreaterThanMaxFee,
    GasPriceLessThanBasefee,
    CallerGasLimitMoreThanBlock,
    CallGasCostMoreThanGasLimit {
        initial_gas: u64,
        gas_limit: u64,
    },
    GasFloorMoreThanGasLimit {
        gas_floor: u64,
        gas_limit: u64,
    },
    RejectCallerWithCode,
    LackOfFundForMaxFee {
        fee: Box<Uint<256, 4>>,
        balance: Box<Uint<256, 4>>,
    },
    OverflowPaymentInTransaction,
    NonceOverflowInTransaction,
    NonceTooHigh {
        tx: u64,
        state: u64,
    },
    NonceTooLow {
        tx: u64,
        state: u64,
    },
    CreateInitCodeSizeLimit,
    InvalidChainId,
    AccessListNotSupported,
    MaxFeePerBlobGasNotSupported,
    BlobVersionedHashesNotSupported,
    BlobGasPriceGreaterThanMax,
    EmptyBlobs,
    BlobCreateTransaction,
    TooManyBlobs {
        max: usize,
        have: usize,
    },
    BlobVersionNotSupported,
    EofCreateShouldHaveToAddress,
    AuthorizationListNotSupported,
    AuthorizationListInvalidFields,
    EmptyAuthorizationList,
    Eip2930NotSupported,
    Eip1559NotSupported,
    Eip4844NotSupported,
    Eip7702NotSupported,
    Eip7873NotSupported,
    Eip7873MissingTarget,

}

Transaction validation error.

Variants

PriorityFeeGreaterThanMaxFee

When using the EIP-1559 fee model introduced in the London upgrade, transactions specify two primary fee fields:

gas_max_fee: The maximum total fee a user is willing to pay, inclusive of both base fee and priority fee.

gas_priority_fee: The extra amount a user is willing to give directly to the miner, often referred to as the โtipโ.

Provided gas_priority_fee exceeds the total gas_max_fee.

GasPriceLessThanBasefee

EIP-1559: gas_price is less than basefee.

CallerGasLimitMoreThanBlock

gas_limit in the tx is bigger than block_gas_limit.

CallGasCostMoreThanGasLimit

Initial gas for a Call is bigger than gas_limit.

Initial gas for a Call contains:

initial stipend gas

gas for access list and input data

Fields

initial_gas: u64

gas_limit: u64

GasFloorMoreThanGasLimit

Gas floor calculated from EIP-7623 Increase calldata cost is more than the gas limit.

Tx data is too large to be executed.

Fields

gas_floor: u64

gas_limit: u64

RejectCallerWithCode

EIP-3607 Reject transactions from senders with deployed code

LackOfFundForMaxFee

Transaction account does not have enough amount of ether to cover transferred value and gas_limit*gas_price.

Fields

fee: Box<Uint<256, 4>>

balance: Box<Uint<256, 4>>

OverflowPaymentInTransaction

Overflow payment in transaction.

NonceOverflowInTransaction

Nonce overflows in transaction.

NonceTooHigh

Fields

tx: u64

state: u64

NonceTooLow

Fields

tx: u64

state: u64

CreateInitCodeSizeLimit

EIP-3860: Limit and meter initcode

InvalidChainId

Transaction chain id does not match the config chain id.

AccessListNotSupported

Access list is not supported for blocks before the Berlin hardfork.

MaxFeePerBlobGasNotSupported

max_fee_per_blob_gas is not supported for blocks before the Cancun hardfork.

BlobVersionedHashesNotSupported

blob_hashes/blob_versioned_hashes is not supported for blocks before the Cancun hardfork.

BlobGasPriceGreaterThanMax

Block blob_gas_price is greater than tx-specified max_fee_per_blob_gas after Cancun.

EmptyBlobs

There should be at least one blob in Blob transaction.

BlobCreateTransaction

Blob transaction canโt be a create transaction.

to must be present

TooManyBlobs

Transaction has more then max blobs

Fields

max: usize

have: usize

BlobVersionNotSupported

Blob transaction contains a versioned hash with an incorrect version

EofCreateShouldHaveToAddress

EOF create should have to address

AuthorizationListNotSupported

EIP-7702 is not enabled.

AuthorizationListInvalidFields

EIP-7702 transaction has invalid fields set.

EmptyAuthorizationList

Empty Authorization List is not allowed.

Eip2930NotSupported

EIP-2930 is not supported.

Eip1559NotSupported

EIP-1559 is not supported.

Eip4844NotSupported

EIP-4844 is not supported.

Eip7702NotSupported

EIP-7702 is not supported.

Eip7873NotSupported

EIP-7873 is not supported.

Eip7873MissingTarget

EIP-7873 initcode transaction should have to address.