revm_context::tx

Struct TxEnv

pub struct TxEnv {

    pub tx_type: u8,
    pub caller: Address,
    pub gas_limit: u64,
    pub gas_price: u128,
    pub kind: TxKind,
    pub value: U256,
    pub data: Bytes,
    pub nonce: u64,
    pub chain_id: Option<u64>,
    pub access_list: AccessList,
    pub gas_priority_fee: Option<u128>,
    pub blob_hashes: Vec<B256>,
    pub max_fee_per_blob_gas: u128,
    pub authorization_list: Vec<Either<SignedAuthorization, RecoveredAuthorization>>,

}

The Transaction Environment is a struct that contains all fields that can be found in all Ethereum transaction, including EIP-4844, EIP-7702, EIP-7873, etc. It implements the Transaction trait, which is used inside the EVM to execute a transaction.

TxEnvBuilder builder is recommended way to create a new TxEnv as it will automatically set the transaction type based on the fields set.

Fields
tx_type: u8

Transaction type

caller: Address

Caller aka Author aka transaction signer

gas_limit: u64

The gas limit of the transaction.

gas_price: u128

The gas price of the transaction.

For EIP-1559 transaction this represent max_gas_fee.

kind: TxKind

The destination of the transaction

value: U256

The value sent to transact_to

data: Bytes

The data of the transaction

nonce: u64

The nonce of the transaction

chain_id: Option<u64>

The chain ID of the transaction

If set to None, no checks are performed.

Incorporated as part of the Spurious Dragon upgrade via EIP-155.

access_list: AccessList

A list of addresses and storage keys that the transaction plans to access

Added in EIP-2930.

gas_priority_fee: Option<u128>

The priority fee per gas

Incorporated as part of the London upgrade via EIP-1559.

blob_hashes: Vec<B256>

The list of blob versioned hashes

Per EIP there should be at least one blob present if max_fee_per_blob_gas is Some.

Incorporated as part of the Cancun upgrade via EIP-4844.

max_fee_per_blob_gas: u128

The max fee per blob gas

Incorporated as part of the Cancun upgrade via EIP-4844.

authorization_list: Vec<Either<SignedAuthorization, RecoveredAuthorization>>

List of authorizations

authorization_list contains the signature that authorizes this caller to place the code to signer account.

Set EOA account code for one transaction via EIP-7702.

Implementations

impl TxEnv

pub fn derive_tx_type(&mut self) -> Result<(), DeriveTxTypeError>

Derives tx type from transaction fields and sets it to tx_type. Returns error in case some fields were not set correctly.

pub fn set_signed_authorization(&mut self, auth: Vec<SignedAuthorization>)

Insert a list of signed authorizations into the authorization list.

pub fn set_recovered_authorization(&mut self, auth: Vec<RecoveredAuthorization>)

Insert a list of recovered authorizations into the authorization list.

impl TxEnv

pub fn builder() -> TxEnvBuilder

Create a new builder for constructing a TxEnv