Crate revm_context

Optimism-specific constants, types, and helpers.

Re-exports

pub use block::BlockEnv;

pub use cfg::CfgEnv;

pub use evm::Evm;

pub use local::LocalContext;

pub use tx::TxEnv;

pub use context::*;

pub use journal::*;

Modules

block

This module contains BlockEnv and it implements Block trait.

cfg

This module contains CfgEnv and implements Cfg trait for it.

context

This module contains Context struct and implements ContextTr trait for it.

either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

evm

This module contains Evm struct.

journal

This module contains Journal struct and implements JournalTr trait for it.

journaled_state

local

Local context that is filled by execution.

result

transaction

tx

This module contains TxEnv struct and implements Transaction trait for it.

Enums

CreateScheme

Create scheme

TransactionType

Transaction types of all Ethereum transaction

Traits

Block

Trait for retrieving block information required for execution.

Cfg

ContextSetters

ContextTr

Trait that defines the context of the EVM execution.

DBErrorMarker

Database error marker is needed to implement From conversion for Error type.

Database

EVM database interface.

JournalTr

Trait that contains database and journal of all changes that were made to the state.

LocalContextTr

Local context used for caching initcode from Initcode transactions.

Transaction

Main Transaction trait that abstracts and specifies all transaction currently supported by Ethereum

Type Aliases

TransactTo

Transaction destination