revm_context::cfg

Trait Cfg

pub trait Cfg {
    type Spec: Into<SpecId> + Clone;

    // Required methods
    fn chain_id(&self) -> u64;
    fn spec(&self) -> Self::Spec;
    fn blob_max_count(&self) -> Option<u64>;
    fn max_code_size(&self) -> usize;
    fn is_eip3607_disabled(&self) -> bool;
    fn is_balance_check_disabled(&self) -> bool;
    fn is_block_gas_limit_disabled(&self) -> bool;
    fn is_nonce_check_disabled(&self) -> bool;
    fn is_base_fee_check_disabled(&self) -> bool;
}

Required Associated Types

type Spec: Into<SpecId> + Clone

Required Methods

fn chain_id(&self) -> u64

fn spec(&self) -> Self::Spec

fn blob_max_count(&self) -> Option<u64>

Returns the blob target and max count for the given spec id. If it is None, check for max count will be skipped.

EIP-7840: Add blob schedule to execution client configuration files

fn max_code_size(&self) -> usize

fn is_eip3607_disabled(&self) -> bool

fn is_balance_check_disabled(&self) -> bool

fn is_block_gas_limit_disabled(&self) -> bool

fn is_nonce_check_disabled(&self) -> bool

fn is_base_fee_check_disabled(&self) -> bool